[pyproject.toml]: Explicitly Specify Package to be Installed
Explicitly specify the `python` directory as "package" that should be
installed by setuptools when installing this project.

This is necessary to avoid the error
`error: Multiple top-level packages discovered in a flat-layout`
when installing this project.

This is some kind of workaround.  Actually, this project does not
provide an installable Python package.  "Installing" this package simply
means to install its dependencies.